defined level class
