+++ assets/js/main.js
@@ -1,13 +1,15 @@
 /* import external classes and instances */
 import k from "./kaboom.js"
-import pickup from "./pickup.js"
+import Level from "./level.js";
+import pickup from "./pickup.js";
+import Tile from "./tile.js";;
 import loadLevel from "./LoadLevel.js";
 
 /* load graphics and other assets */
 loadSprite("santa", "./assets/sprites/santa.png");
 loadSprite("enemy", "./assets/sprites/enemy.png");
 
-const tile = loadSprite("tile", "./assets/sprites/tile.png");
+const tile = loadSprite("tile", "./assets/tiles/ground-substrate.png");
 
 /* define movement speed and jump height for santa
 load the sprite and physics properties */
@@ -27,49 +29,12 @@ const santa = add([
 
 loadSound("jump", "./assets/sfx/jump.wav");
 
-/* call loadlevel function and deposit
-the level data into a variable */
-const level = await loadLevel();
-const levelData = level.layers[0].data;
-
-/* explicitly declare level constants
-so we can use them for loading the level */
-const TILE_WIDTH = 24;
-const TILE_HEIGHT = 24;
-const TILES_ACROSS = 135;
-
-/* row and column are used during the level load procedure
-to determine the exact placement of each tile */
-let row = -1;
-let column = -1;
-
-/* tile object, x,y position as well as its type ie. graphical style */
-class Tile {
-    constructor(xPos, yPos) {
-        this.x = xPos;
-        this.y = yPos;
-        this.type = 1;
-    }
-}
+let levelFile = await loadLevel('level1');
+let currentLevel = new Level(levelFile);
 
 /* These are the tiles in each level data structure that are not
 equal to 0, ie. tiles that are visible */
-const visibleTiles = new Set();
-
-/* cycle through the loaded level data and for each visible tile,
-add it to the the visibleTiles set for rendering */
-levelData.forEach(tile => {
-    column += 1;
-
-    if (tile > 0) {
-        visibleTiles.add(new Tile(column * TILE_WIDTH, row * TILE_HEIGHT))
-    }
-
-    if (column === TILES_ACROSS) {
-        column = 0;
-        row += 1;
-    }
-});
+let visibleTiles = currentLevel.getVisibleTiles;
 
 // camera follows player
 santa.onUpdate(() => {
